use std lib for http and zip
